% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{carsimr}
\alias{carsimr}
\title{Create a carsimr object}
\usage{
carsimr(grid)
}
\arguments{
\item{grid}{A matrix representing the grid or an existing carsimr object.}
}
\value{
A validated carsimr object.
}
\description{
This function creates a new carsimr object with the specified grid and
validates it.
}
