% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_grid.R
\name{initialize_grid}
\alias{initialize_grid}
\title{Initialize Grid with Moving Cars}
\usage{
initialize_grid(rho, dims, prob_blue)
}
\arguments{
\item{rho}{The number of grid cells that should be filled with cars. It can
be either a proportion or an integer.}

\item{dims}{Size of the grid or dimensions of the matrix as a numeric vector
(e.g., c(rows, cols)).}

\item{prob_blue}{Probability of selecting a blue car. Should be between 0 and
1.}
}
\value{
Initialized grid as a carsimr class object.
}
\description{
This function creates a grid with moving cars based on specified parameters.
}
\examples{
# Initialize grid with proportion of cars
initialize_grid(0.3, c(10, 10), 0.5)
# Initialize grid with specific number of cars
initialize_grid(20, c(15, 15), 0.6)
}
