% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialize_grid_cpp}
\alias{initialize_grid_cpp}
\title{Initialize a Grid with Cars}
\usage{
initialize_grid_cpp(rho, dims, prob_blue)
}
\arguments{
\item{rho}{A numeric value that can either be a proportion of cells to fill
(between 0 and 1) or a total number of cars if greater than 1.}

\item{dims}{An integer vector specifying the dimensions of the grid (rows,
columns).}

\item{prob_blue}{A numeric probability (between 0 and 1) that a car will be
blue;
the probability of a car being red is 1 - prob_blue.}
}
\value{
A numeric matrix representing the grid, where 0 indicates an empty
cell, 1 indicates a blue car, and 2 indicates a red car.
}
\description{
This function creates a grid of the specified dimensions where each cell
may randomly contain a car based on the specified probability. The type
of car (blue or red) is also determined based on a probability.
}
