% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_functions.R
\name{move_blue}
\alias{move_blue}
\title{Move Blue Cars}
\usage{
move_blue(t_grid)
}
\arguments{
\item{t_grid}{Grid from the previous step.}
}
\value{
Updated grid with moved blue cars as a carsimr class object.
}
\description{
This function moves blue cars upwards by one grid cell if the cell above is
empty.
}
