% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_functions.R
\name{move_cars}
\alias{move_cars}
\title{Move Cars}
\usage{
move_cars(t_grid, trials)
}
\arguments{
\item{t_grid}{A grid of cars of class carsimr.}

\item{trials}{Number of car movements to simulate.}
}
\value{
An object of class carsimr_list containing the initial grid and each
intermediate grid after each trial.
}
\description{
This function simulates the movement of cars on the grid for a specified
number of trials.
The red and blue cars take turns moving.
}
\examples{
# Move cars for 10 trials
move_cars(initialize_grid(0.3, c(10, 10), 0.5), 10)
}
