% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{move_cars_cpp}
\alias{move_cars_cpp}
\title{Move Cars on a Grid}
\usage{
move_cars_cpp(grid, trials)
}
\arguments{
\item{grid}{A numeric matrix representing the initial state of the grid,
where 0 indicates an empty cell, 1 indicates a blue car, and 2 indicates a red car.}

\item{trials}{An integer representing the number of movements to simulate.}
}
\value{
A numeric matrix representing the grid after all movements have been simulated.
}
\description{
This function simulates moving cars in a grid for a specified number of trials.
Cars move according to specific rules: red cars move right on even-numbered trials
and blue cars move up on odd-numbered trials.
}
