% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_functions.R
\name{move_red}
\alias{move_red}
\title{Move Red Cars}
\usage{
move_red(t_grid)
}
\arguments{
\item{t_grid}{Grid from the previous step.}
}
\value{
Updated grid with moved red cars as a carsimr class object.
}
\description{
This function moves red cars to the right by one grid cell if the cell to the
right is empty.
}
