% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_carsimr}
\alias{new_carsimr}
\title{Create a new carsimr objects}
\usage{
new_carsimr(grid = matrix(0, nrow = 1, ncol = 1))
}
\arguments{
\item{grid}{A matrix representing the grid or an existing carsimr object.}
}
\value{
A carsimr object.
}
\description{
This function creates a new carsimr object with the specified grid.
}
