% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_carsimr_list}
\alias{new_carsimr_list}
\title{Create a new carsimr_list object}
\usage{
new_carsimr_list(grid_list)
}
\arguments{
\item{grid_list}{A list of grid matrices.}
}
\value{
A carsimr_list object.
}
\description{
This function creates a new carsimr_list object with the specified list of
grids.
}
