% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.carsimr}
\alias{plot.carsimr}
\title{Plot functions}
\usage{
\method{plot}{carsimr}(x, ...)

\method{plot}{carsimr}(x, ...)
}
\arguments{
\item{x}{An object of class 'carsimr'}

\item{...}{Additional arguments}
}
\value{
A plot of the 'carsimr' grid
}
\description{
These functions provide plotting capabilities for objects of class 'carsimr'
and 'carsimr_list'.
}
