% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.carsimr_list}
\alias{plot.carsimr_list}
\title{Plot Functions for carsimr_list Objects}
\usage{
\method{plot}{carsimr_list}(x, pause = 0.9, ...)

\method{plot}{carsimr_list}(x, pause = 0.9, ...)
}
\arguments{
\item{x}{An object of class 'carsimr_list'}

\item{pause}{Value to pause the plots}

\item{...}{Additional arguments}
}
\value{
Plotting of carsimr_list object with a specified pause.
}
\description{
Plot Functions for carsimr_list Objects

Plot method for carsimr_list objects
}
