% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{validate_carsimr}
\alias{validate_carsimr}
\title{Validate carsimr object}
\usage{
validate_carsimr(x)
}
\arguments{
\item{x}{An object to be validated.}
}
\value{
The input object if it is valid.
}
\description{
This function validates whether the input object is a valid carsimr object.
}
