% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{validate_carsimr_list}
\alias{validate_carsimr_list}
\title{Validate carsimr_list object}
\usage{
validate_carsimr_list(x)
}
\arguments{
\item{x}{An object to be validated.}
}
\value{
The input object if it is valid.
}
\description{
This function validates whether the input object is a valid carsimr_list
object.
}
