% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{validate_dims}
\alias{validate_dims}
\title{Validate dimensions of the grid}
\usage{
validate_dims(dims)
}
\arguments{
\item{dims}{Numeric vector specifying the dimensions of the grid.}
}
\value{
The validated dimensions vector.
}
\description{
Checks that dimensions provided are a numeric vector of length 2 and each
element is a positive integer.
}
