% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{validate_move_cars_inputs}
\alias{validate_move_cars_inputs}
\title{Validate Inputs for Moving Cars Rcpp}
\usage{
validate_move_cars_inputs(grid, trials)
}
\arguments{
\item{grid}{A matrix representing the grid of cars, where elements should be
either 0, 1, or 2.}

\item{trials}{An integer for the number of trials or movements to simulate.}
}
\value{
Invisible NULL, this function is used purely for validation and will
}
\description{
This function checks the validity of the inputs for the move_cars_cpp
function.
}
