% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{validate_prob_blue}
\alias{validate_prob_blue}
\title{Validate probability of selecting a blue car}
\usage{
validate_prob_blue(prob_blue)
}
\arguments{
\item{prob_blue}{Numeric value showing the probability of picking a blue car.}
}
\value{
The validated prob_blue value.
}
\description{
Checks that prob_blue is a numeric value between 0 and 1.
}
