% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{validate_rho}
\alias{validate_rho}
\title{Validate and adjust rho parameter}
\usage{
validate_rho(rho)
}
\arguments{
\item{rho}{Numeric value showing the proportion or absolute number of cars.}
}
\value{
The validated and possibly adjusted rho value.
}
\description{
Ensures that rho is a numeric value. If rho is greater than 1, it rounds rho
to the nearest integer, assuming it represents an absolute number of cars
rather than a proportion.
}
